+++ base-demo-02-xtgl/src/main/java/com/basedemo02xtgl/basedemo02xtgl/entity/SysRoleMenu.java
@@ -15,9 +15,9 @@ public class SysRoleMenu implements Serializable {
     private static final long serialVersionUID = 1L;
 
 
-    private Long id;
-    private Long roleId;
-    private Long menuId;
+    private Integer id;
+    private Integer roleId;
+    private Integer menuId;
 
 
 }
